Fixed some warnings, and added ability to run 'protected code' - that is, code that catches system-level exceptions and turns them into software exceptions (which now guards graphics and audio rendering). Fixed some bugs here and there in window functions, added a date function, fixed presets, added ability to append to exclusive files.

+++ CPLTests.cpp
@@ -202,7 +202,7 @@ namespace cpl
 
 						Misc::PreciseDelay(msPerRender);
 					}
-					drops = stream.getPerfMeasures().droppedAudioFrames;
+					drops = static_cast<std::size_t>(stream.getPerfMeasures().droppedAudioFrames);
 				}
 			);
